+++ xsolla-store-sdk/src/main/java/com/xsolla/android/store/entity/request/inventory/ConsumeItemBody.java
@@ -5,11 +5,11 @@ import com.google.gson.annotations.SerializedName;
 public class ConsumeItemBody {
 
     private String sku;
-    private int quantity;
+    private long quantity;
     @SerializedName("instance_id")
     private String instanceId;
 
-    public ConsumeItemBody(String sku, int quantity, String instanceId) {
+    public ConsumeItemBody(String sku, long quantity, String instanceId) {
         this.sku = sku;
         this.quantity = quantity;
         this.instanceId = instanceId;
